refactor: split `IntersectionObserver` into `InView` and `InViewContextProvider` (MET-1579) (#29)
* refactor: split `IntersectionObserver` into `InView` and `InViewContextProvider`

First, the `InView` name is a bit more accurate, as it says whether the child element is in view, and it resembles what is exported by the https://www.npmjs.com/package/react-intersection-observer library.

Second, I still can't decide what will be better - exporting `inView` as a context (and thus making the parent component rerender on any context changes), or instead, to avoid rerenders, just have two react nodes (children and fallback) and render just one of them, depending on whether element is in view. Let's do both for now.

* fixup! refactor: split `IntersectionObserver` into `InView` and `InViewContextProvider`